# openapi/openapi.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: dManga API
  description: >-
    API schema of backend server for dManga
servers:
  - url: 'http://localhost:8080'
paths:
  /users/{wallet_address}:
    get:
      description: >
        Return a user
      operationId: findUser
      parameters:
        - name: wallet_address
          in: path
          description: wallet_address of user to find
          required: true
          schema:
            type: string
      responses:
        '200':
          description: user response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/User'
        '404':
          description: user not found
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/User'

        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      description: >
        Create or update a user
      operationId: putUser
      parameters:
        - name: wallet_address
          in: path
          description: wallet_address of user to find
          required: true
          schema:
            type: string
      requestBody:
        description: user info to register
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - "name"
              property:
                name:
                  type: "string"
                  example: "Taro"
                picture:
                  type: "string"
                  example: "https://theme.zdassets.com/theme_assets/2313093/fe875eef30ee9649f253e8188f3438bb3be27e96.png"
                biography:
                  type: "string"
                  example: "こんにちは、よろしくおねがいします。"

      responses:
        '200':
          description: user response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    delete:
      description: >
        Delete a user
      operationId: deleteUser
      parameters:
        - name: wallet_address
          in: path
          description: wallet_address of user to delete
          required: true
          schema:
            type: string
      responses:
        '200':
          description: user response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
      
components:
  schemas:
    User:
      type: "object"
      required:
        - "wallet_address"
      properties:
        wallet_address:
          type: "string"
          example: "0xc886E08423C65b2B28AAD7CCF2b87F973C915736"
        name:
          type: "string"
          example: "Taro"
        picture:
          type: "string"
          example: "https://theme.zdassets.com/theme_assets/2313093/fe875eef30ee9649f253e8188f3438bb3be27e96.png"
        biography:
          type: "string"
          example: "こんにちは、よろしくおねがいします。"
        
    Error:
      type: "object"
      properties:
        code:
          type: "integer"
          format: "int32"
        type:
          type: "string"
        message:
          type: "string"
externalDocs:
  description: "Find out more about Swagger"
  url: "http://swagger.io"